+++ queueMessages.js
@@ -58,8 +58,8 @@ register('chat', (cost, item, recipient, event) => {
         if (attemptMatchObject) {   
             storedClaimedMessages.pop(idx);
             const shownCollector = attemptMatchObject.collector.removeFormatting().trim() === Player.getName() ? '' : `${attemptMatchObject.collector} `;
-            replaceAuctionMessage(event, `${AH_PREFIX}${shownCollector}&6COLLECTED: ${attemptMatchObject.item} &7for &6${attemptMatchObject.cost} &7to ${attemptMatchObject.buyer}`);    
-            return;
+            replaceAuctionMessage(event, `${AH_PREFIX}${shownCollector}&6CLAIMED: ${attemptMatchObject.item} &7for &6${attemptMatchObject.cost} &7to ${attemptMatchObject.buyer}`);    
+            return;         
         }
     }  
     storedClaimedMessages.push(message);
@@ -73,7 +73,7 @@ register('chat', (collector, coins, event) => {
     const collectorName = stripRank(collector).trim();
     if (collectorName !== Player.getName()) {   
         const matchObject = getAhMessageInfo(message);  
-        replaceAuctionMessage(event, `${AH_PREFIX}COLLECTED: &6${matchObject.cost} &7by ${matchObject.collector}&7!`);
+        replaceAuctionMessage(event, `${AH_PREFIX}CLAIMED: &6${matchObject.cost} &7by ${matchObject.collector}&7!`);
         return;
     }
 
@@ -83,7 +83,7 @@ register('chat', (collector, coins, event) => {
         if (attemptMatchObject) {   
             storedClaimedMessages.pop(idx);
             const shownCollector = attemptMatchObject.collector.removeFormatting().trim() === Player.getName() ? '' : `${attemptMatchObject.collector} `;
-            replaceAuctionMessage(event, `${AH_PREFIX}${shownCollector}&6COLLECTED: ${attemptMatchObject.item} &7for &6${attemptMatchObject.cost} &7to ${attemptMatchObject.buyer}`);  
+            replaceAuctionMessage(event, `${AH_PREFIX}${shownCollector}&6CLAIMED: ${attemptMatchObject.item} &7for &6${attemptMatchObject.cost} &7to ${attemptMatchObject.buyer}`);  
             return;
         }
     }   
@@ -91,12 +91,81 @@ register('chat', (collector, coins, event) => {
     cancel(event);
 }).setCriteria('${collector} collected an auction for ${coins}');
 
-// let storedExpiredMessages = [];
-// register('chat', (item, event) => {
-//     if (!getInSkyblock()) return;
-// }).setCriteria('You claimed ${item} back from your expired auction!');
+function getExpiredInfo(msg) {
+    let msgType = msg.includes('You claimed') ? "personal" : "coop";
+    let resultItem, resultCollector;
+    if (msgType === "personal") {
+        const personalRegex = /&r&eYou claimed &r&f&r(.+) &r&eback from your expired auction!&r/;
+        const match = msg.match(personalRegex);
+        if (match) {
+            const [_, formattedItem] = match;
+            resultItem = formattedItem;
+        }
 
-// register('chat', (collector, event) => {
-//     if (!getInSkyblock()) return;
-// }).setCriteria('${collector} collected an expired auction!');
+    } else if (msgType === "coop") {
+        const coopRegex = /(&[a-qs-z0-9])(.+)&r&f &r&ecollected an expired auction!&r/;
+        const match = msg.match(coopRegex);
+        if (match) {
+            const [_, collectorColor, collectorName] = match;
+            const formattedCollector = `${collectorColor}${stripRank(collectorName.removeFormatting())}`;
+            resultCollector = formattedCollector;
+        }
+    }
+
+    return {
+        type: msgType,
+        item: resultItem,
+        collector: resultCollector
+    }
+};
+
+function attemptExpiredMatch(msg1, msg2) {
+    const msg1Info = getExpiredInfo(msg1);
+    const msg2Info = getExpiredInfo(msg2);
+    if (msg1Info.type === msg2Info.type) return null;
+    return {
+        item: msg1Info.item ?? msg2Info.item,
+        collector: msg1Info.collector ?? msg2Info.collector
+    };
+}
+
+let storedExpiredMessages = [];
+register('chat', (item, event) => {
+    if (!getInSkyblock()) return;
+    const message = ChatLib.getChatMessage(event, true);
+    for (let idx = 0; idx < storedExpiredMessages.length; idx++) {
+        let expiredMessage = storedExpiredMessages[idx];
+        const attemptMatchObject = attemptExpiredMatch(message, expiredMessage);
+        if (attemptMatchObject) {
+            storedExpiredMessages.pop(idx);
+            replaceAuctionMessage(event, `${AH_PREFIX}${attemptMatchObject.collector} &6CLAIMED &cEXPIRED: &r${attemptMatchObject.item}&7!`);
+            return;
+        }
+    }
+    storedExpiredMessages.push(message);
+    cancel(event);
+}).setCriteria('You claimed ${item} back from your expired auction!');
+
+register('chat', (collector, event) => {
+    if (!getInSkyblock()) return;
+    if (collector === 'You') return;
+    const message = ChatLib.getChatMessage(event, true);
+    const collectorName = stripRank(collector).trim();
+    if (collectorName !== Player.getName()) {
+        const matchObject = getExpiredInfo(message);
+        replaceAuctionMessage(event, `${AH_PREFIX}&r${matchObject.collector} &7claimed an &cexpired&7 item!`);
+    }
+    
+    for (let idx = 0; idx < storedExpiredMessages.length; idx++) {
+        let expiredMessage = storedExpiredMessages[idx];
+        const attemptMatchObject = attemptExpiredMatch(message, expiredMessage);
+        if (attemptMatchObject) {   
+            storedExpiredMessages.pop(idx);
+            replaceAuctionMessage(event, `${AH_PREFIX}${attemptMatchObject.collector} &6CLAIMED &cEXPIRED: &r${attemptMatchObject.item}&7!`);
+            return;
+        }
+    }
+    storedExpiredMessages.push(message);
+    cancel(event);
+}).setCriteria('${collector} collected an expired auction!');