+++ changelog.md
@@ -15,5 +15,5 @@
 + added firstInstall
 + changed gameload message
 + added error "this auction couldn't be found!"
-+ WIP: refactored and reformatted claim auction messages
++ refactored and reformatted claim auction messages
 + WIP: refactored and reformatted claim expired auction messages